feat: move all images into images folder
