+++ src/components/layout/header.tsx
@@ -2,7 +2,7 @@
 
 import Image from "next/image";
 import Link from "next/link";
-import logo from "@public/logo.svg";
+import logo from "@public/images/logos/xplorer_logo.svg";
 import { usePathname } from "next/navigation";
 import clsx from "clsx";
 import { menus } from "./helper";
